register user

+++ base/views.py
@@ -6,12 +6,21 @@ from .models import Room, Topic
 from .forms import RoomForm
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 
 
 
 def loginPage(request):
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -26,14 +35,26 @@ def loginPage(request):
         else:
             messages.error(request, "Username or password does not exist.")
 
-    context={}
+    context={"page": page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
 
-
+def registerUser(request):
+    page = 'register'
+    form = UserCreationForm()
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occurred during registration.")
 
 
 def home(request):
@@ -54,6 +75,7 @@ def rooms(request,pk):
     context={"room": room}
     return render(request, 'base/rooms.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -64,9 +86,13 @@ def createRoom(request):
     context={"form": form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance = room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -77,8 +103,11 @@ def updateRoom(request,pk):
     context={'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id = pk)
+    if request.user != room.host:
+            return HttpResponse("You are not allowed here!!")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
